vethsetup: do not touch originally created testeth0 connection

+++ prepare/vethsetup.sh
@@ -186,10 +186,11 @@ function setup_veth_env ()
     nmcli c modify testeth0 ipv4.route-metric 99 ipv6.route-metric 99
     nmcli c u testeth0
 
-    # THIS NEED TO BE DONE HERE FOR RECREATION REASONS
-    # Copy final connection to /tmp/testeth0 for later in test usage
-    yes 2>/dev/null | cp -rf /etc/sysconfig/network-scripts/ifcfg-testeth0 /tmp/testeth0
-
+    if [ ! -e /tmp/testeth0 ] ; then
+        # THIS NEED TO BE DONE HERE FOR RECREATION REASONS
+        # Copy final connection to /tmp/testeth0 for later in test usage
+        yes 2>/dev/null | cp -rf /etc/sysconfig/network-scripts/ifcfg-testeth0 /tmp/testeth0
+    fi
 
     # On s390x sometimes this extra default profile gets created in addition to custom static original one
     # Let's get rid of that
@@ -343,7 +344,6 @@ function teardown_veth_env ()
     sleep 1
     nmcli con reload
     sleep 1
-    rm /tmp/testeth0
 
     # Rename the device back to ORIGNAME
     if [ "$ORIGDEV" != "eth0" ]; then
